+++ src/server/network/protocol/protocolgame.cpp
@@ -1382,7 +1382,7 @@ void ProtocolGame::parsePacketFromDispatcher(NetworkMessage &msg, uint8_t recvby
 
 		default:
 			std::string hexString = fmt::format("0x{:02x}", recvbyte);
-			g_logger().warn("Player '{}' sent unknown packet header: hex[{}], decimal[{}]", player->getName(), asUpperCaseString(hexString), recvbyte);
+			g_logger().debug("Player '{}' sent unknown packet header: hex[{}], decimal[{}]", player->getName(), asUpperCaseString(hexString), recvbyte);
 			break;
 	}
 }