+++ markdowns/CHANGELOG.md
@@ -8,6 +8,7 @@
 - New tag for rune spells: `rune:removeOnUse(bool)`. This can be used to specify a particular rune to remove charges, even when `removeChargesFromRunes` is set to false in `config.lua`. ([Tryller](https://github.com/jprzimba))
 - Moved files from data-global/migrations to data/migrations. ([Tryller](https://github.com/jprzimba))
 - New configurations in `config.lua`: `timeToRegenMinuteStamina` and `timeToRegenMinutePremiumStamina`. ([Tryller](https://github.com/jprzimba))
+- Added a new attribute `attackspeed` to mounts.xml, allowing mounts to increase player attack speed. ([Tryller](https://github.com/jprzimba))
 
 ## Added files
 
@@ -20,6 +21,7 @@
 - data/scripts/creaturescripts/player/login.lua
 - data/scripts/creaturescripts/player/regenerate_stamina.lua
 - data/XML/groups.xml
+- data/XML/mounts.xml
 
 ## Deleted files