+++ markdowns/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Protocol 14.10 support. ([Tryller](https://github.com/jprzimba))
 - New protocol 14.10 mounts and outfits. ([Tryller](https://github.com/jprzimba))
 - New protocol 14.10 assets. ([Tryller](https://github.com/jprzimba))
+- New stash for 14.10 protocol. ([Tryller](https://github.com/jprzimba))
 - Load Gamemaster, Community Manager and God outfits from groups.xml ([Tryller](https://github.com/jprzimba)).
 - New tag for rune spells: `rune:removeOnUse(bool)`. This can be used to specify a particular rune to remove charges, even when `removeChargesFromRunes` is set to false in `config.lua`. ([Tryller](https://github.com/jprzimba))
 - Moved files from data-global/migrations to data/migrations. ([Tryller](https://github.com/jprzimba))